requirements-html.txt
# HTML E-Laws Ingestion Pipeline Requirements

# LangChain and document processing
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# HTML parsing
beautifulsoup4>=4.12.0
requests>=2.31.0

# LLM and embeddings
openai>=1.3.0
sentence-transformers>=2.2.0

# Database
neo4j>=5.14.0
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
